requirements.txt
torch
torchvision
tensorboard
numpy
matplotlib
Pillow
